+++ src/docs/devguide/hugo/site/content/release-notes.md
@@ -1,5 +1,12 @@
 This page describes changes to the Money Transfer Retailer Interface implemented across different releases of the interface.
 
+## v2.18.0
+
+Released xx Dec 2019
+
+- Updated the `MoneyTransferAuthResponse` and `MoneyTransferAuthRequest` models with a new `amounts` field:
+    - The `amounts` field allows for more amounts to be used to describe a Money Transfer transaction.
+
 ## v2.17.0
 
 Released 06 November 2019